app/requirements.txt
Django==4.2.2
python-dotenv
psycopg==3.1.8
django-monthfield